[test] Remove global `styled-components` install
Checking some assumptions:

It's only used for running `pnpm next <test-dir>`
locally. However, it doesn't necessarily reflect
the actual `styled-components` version of the test
project in question so it's a footgun.

It also slows down all installs which is not negligible
given the large package graph we could remove in
this PR.

If you want to run a test project locally that requires
some additional packages, use `pnpm --ignore-workspace --no-lockfile install`
instead.
For tests that don't have a dedicated package.json,
use `pnpm --ignore-workspace --no-lockfile add styled-components@<version>.`.

A dedicated package.json for each fixture should be
preferred for ergonomics.
It's still duplicated but at least the tested version
in `dependencies` is closer to the one in the `package.json`.
Having a single version in the monorepo is too far
away causing drift.